Fixes #631: session basedd login plugin
