+++ mapcomposer/app/applications/example/templates/example.html
@@ -263,6 +263,7 @@
                     portalConfig: serverConfig.portalConfig,
 					loadingPanel: serverConfig.loadingPanel || {},
 					geoStoreBaseURL: serverConfig.geoStoreBase || ('http://' + window.location.host + '/geostore/rest/'),	 
+                    adminUrl: serverConfig.adminUrl || ('http://' + window.location.host + '/admin/'),
                     about: {
                         title: "Custom Map",
                         'abstract': "Custom Map",
