+++ mapcomposer/app/config.js
@@ -5,6 +5,7 @@ var urls = [
     [(/^\/(index(.html)?)?/), require("./root/index").app],
     //[(/^\/(login)/), require("./root/login").app],
     [(/^\/(loginpage)/), require("./root/loginpage").app],
+    [(/^\/(session)/), require("./root/session").app],
     //[(/^\/(maps(\/\d+)?)/), require("./root/maps").app],
 	//[(/^\/(geonetwork)/), require("./root/geonetwork").app],  // Enable this only for the GeoNetwork integration
     [(/^\/(composer)/), require("./root/composer").app],
