+++ mapcomposer/app/static/externals/gxp/src/script/plugins/SaveDefaultContext.js
@@ -67,10 +67,10 @@ gxp.plugins.SaveDefaultContext = Ext.extend(gxp.plugins.Tool, {
      */
     addResourceButtonText: "Add Map",
     
-    /** api: config[auth]
+    /** api: config[authHeader]
      *  ``String``
      */
-    auth: null,
+    authHeader: null,
 	
     /**
     * Property: contextMsg
@@ -128,6 +128,18 @@ gxp.plugins.SaveDefaultContext = Ext.extend(gxp.plugins.Tool, {
     */
 	mapDescriptionLabel: "Description",
 	
+    getUser: function() {
+        if(this.target.userDetails && this.target.userDetails.user) {
+            return this.target.userDetails.user.name;
+        } else {
+            var auth = this.getAuth();
+            if(auth) {
+                return Base64.decode(auth.split(' ')[1]);
+            }
+        }
+        return '';
+    },
+    
 	/**
 	 * Property: conflictErrMsg
 	 * {string}
@@ -149,7 +161,7 @@ gxp.plugins.SaveDefaultContext = Ext.extend(gxp.plugins.Tool, {
             tooltip: this.saveDefaultContextActionTip,
             handler: function() {	
                     
-				  if(this.target.auth){
+				  if(this.target.authHeader){
                       
 					  var configStr = Ext.util.JSON.encode(this.target.getState()); 
 					  
@@ -183,7 +195,7 @@ gxp.plugins.SaveDefaultContext = Ext.extend(gxp.plugins.Tool, {
 									  
 						  loginWin.destroy();
 						  
-						  thisObj.auth = 'Basic ' + Base64.encode(user + ':' + pass);           
+						  thisObj.authHeader = 'Basic ' + Base64.encode(user + ':' + pass);           
 						  var configStr = Ext.util.JSON.encode(thisObj.target.getState()); 
 						  
 						  if(thisObj.target.mapId == -1){
@@ -200,7 +212,7 @@ gxp.plugins.SaveDefaultContext = Ext.extend(gxp.plugins.Tool, {
 							  var method = 'PUT';
 							  var contentType = 'application/json';
 							  
-							  thisObj.save(url, method, contentType, configStr, thisObj.auth);
+							  thisObj.save(url, method, contentType, configStr, thisObj.authHeader);
 						  }
 					  };
 					  
@@ -270,10 +282,10 @@ gxp.plugins.SaveDefaultContext = Ext.extend(gxp.plugins.Tool, {
               this.target.modified = false;
               
 			  //
-			  // if the user change language the page is reloaded and this.auth is cleared
+			  // if the user change language the page is reloaded and this.authHeader is cleared
 			  //
-			  if(!this.auth)
-				this.auth = auth;
+			  if(!this.authHeader)
+				this.authHeader = auth;
 				
               this.target.mapId = response.responseText;
               
@@ -301,7 +313,7 @@ gxp.plugins.SaveDefaultContext = Ext.extend(gxp.plugins.Tool, {
            },
            failure:  function(response, opts){
               mask.hide();
-			  this.auth = null;
+			  this.authHeader = null;
 			  
 			  
 			  
@@ -396,7 +408,8 @@ gxp.plugins.SaveDefaultContext = Ext.extend(gxp.plugins.Tool, {
                             var mapName = Ext.getCmp("diag-text-field").getValue();        
                             var mapDescription = Ext.getCmp("diag-text-description").getValue(); 
                             var auth = plugin.getAuth();
-							var owner = Base64.decode(auth.split(' ')[1]);
+                            var owner = plugin.getUser();
+							//var owner = Base64.decode(auth.split(' ')[1]);
 							owner = owner.split(':')[0];
                             var resourceXML = 
 								'<Resource>' +
