+++ mapcomposer/app/templates/embedded.html
@@ -228,6 +228,7 @@
 					portalConfig: serverConfig.portalConfig,
 					loadingPanel: serverConfig.loadingPanel || {},
 					geoStoreBaseURL: serverConfig.geoStoreBase || ('http://' + window.location.host + '/geostore/rest/'),	 
+                    adminUrl: serverConfig.adminUrl || ('http://' + window.location.host + '/admin/'),
                     about: {
                         title: "Custom Map",
                         'abstract': "Custom Map",
