+++ mapcomposer/app/templates/geonetwork.html
@@ -413,6 +413,7 @@
 				tools: serverConfig.tools,
                 removeTools: serverConfig.removeTools,
                 geoStoreBaseURL: serverConfig.geoStoreBase || ('http://' + window.location.host + '/geostore/rest/'),
+                adminUrl: serverConfig.adminUrl || ('http://' + window.location.host + '/admin/'),
                 portalConfig:serverConfig.portalConfig,
 				loadingPanel: serverConfig.loadingPanel || {},
                 defaultSourceType: "gxp_wmssource",
