+++ mapcomposer/app/templates/test/test.html
@@ -319,6 +319,7 @@
                 proxy: proxy,
 				tools: serverConfig.tools,
 				geoStoreBaseURL: serverConfig.geoStoreBase || ('http://' + window.location.host + '/geostore/rest/'),
+                adminUrl: serverConfig.adminUrl || ('http://' + window.location.host + '/admin/'),
                 portalConfig: serverConfig.portalConfig,
 				loadingPanel: serverConfig.loadingPanel || {},
                 defaultSourceType: "gxp_wmssource",
