+++ mapcomposer/app/templates/viewer.html
@@ -261,6 +261,7 @@
                     portalConfig: serverConfig.portalConfig,
 					loadingPanel: serverConfig.loadingPanel || {},
 					geoStoreBaseURL: serverConfig.geoStoreBase || ('http://' + window.location.host + '/geostore/rest/'),	 
+                    adminUrl: serverConfig.adminUrl || ('http://' + window.location.host + '/admin/'),
                     about: {
                         title: "Custom Map",
                         'abstract': "Custom Map",
