+++ mapcomposer/buildjs.cfg
@@ -269,7 +269,9 @@ include =
     plugins/MetadataExplorer.js
 	plugins/MousePosition.js  	
 	plugins/GeoLocationMenu.js
+    plugins/Login.js
     plugins/GeoStoreLogin.js
+    plugins/SessionLogin.js
     plugins/GeoStoreAccount.js
 	plugins/WMSFilterView.js
 	plugins/FeatureEditorGrid.js
